Move handlers to a new package
