+++ WordPress/src/main/java/org/wordpress/android/ui/deeplinks/handlers/DeepLinkHandler.kt
@@ -1,4 +1,4 @@
-package org.wordpress.android.ui.deeplinks
+package org.wordpress.android.ui.deeplinks.handlers
 
 import androidx.lifecycle.LiveData
 import org.wordpress.android.ui.deeplinks.DeepLinkNavigator.NavigateAction
