+++ WordPress/src/main/java/org/wordpress/android/ui/deeplinks/handlers/NotificationsLinkHandler.kt
@@ -1,7 +1,8 @@
-package org.wordpress.android.ui.deeplinks
+package org.wordpress.android.ui.deeplinks.handlers
 
 import org.wordpress.android.ui.deeplinks.DeepLinkNavigator.NavigateAction
 import org.wordpress.android.ui.deeplinks.DeepLinkNavigator.NavigateAction.OpenNotifications
+import org.wordpress.android.ui.deeplinks.DeepLinkingIntentReceiverViewModel
 import org.wordpress.android.util.UriWrapper
 import javax.inject.Inject
