+++ WordPress/src/main/java/org/wordpress/android/ui/deeplinks/handlers/ServerTrackingHandler.kt
@@ -1,4 +1,4 @@
-package org.wordpress.android.ui.deeplinks
+package org.wordpress.android.ui.deeplinks.handlers
 
 import com.android.volley.Request.Method
 import com.android.volley.RequestQueue
